+++ CHANGELOG.rst
@@ -5,6 +5,9 @@ Changelog for Metview's Python interface
 1.7.0
 ------------------
 - added animate() function for animated plots in Jupyter notebooks
+- allowed cfgrib backend keyword arguments to be passed to Fieldset.to_dataset()
+- Fieldset out-of-range indexing now raises an IndexError
+
 
 1.6.1
 ------------------